Add files via upload
Complete the code in the editor below. The variables, , and , are already declared and initialized for you. You should:

Declare variables: one of type int , one of type double , and one of type String .
Read lines of input from stdin (according to the sequence given in the Input Format section below) and initialize your variables.
Use the operator to perform the following operations:
Print the sum of your int variable on a new line.
Print the sum of your double variable to a scale of one decimal place on a new line.
Concatenate the string you read as input and print the result on a new line.
Note: If you are using a language that does not support string concatenation (e.g., C), you can simply print one variable immediately after the other on the same line. The string you entered in your editor should be printed first, immediately followed by the string you read as input.

Input Format

The first line contains an integer that you should add.
The second line contains a double that you must add to.
The third line contains a string that you must concatenate to.
Output format

Print the sum of both integers on the first line, the sum of both doubles (scaled to decimal places) on the second line, and then the two concatenated strings on the third line.